Transfer code to new separate module
